[backend] Fix nil chan
Game channel was not initialized in the beginning of the game. Thus there
was a race condition between initializing the channel and trying to read
from it in the select. Since reading from a nil channel is blocking one
of the players was stuck in the select.

Remove some prints.