[backend] Separate player word model.
Since we need to be able to add only words for a specific user but not
the id of the user who had guessed it in a game, separate the playerWord
model into two - userDictionary and gameWord. gameWord holds an id for a
userDictionary the id of the plater who guessed it and the game id.