[backend] Remove database user from game
Create a separate struct for user that has to be sent to the cleint.
The User struct is sent to the client.
The LoginUser struct is used when verifying login/register.

Rename Players -> Words since it now holds only words information.
Separate the players' ids and data into separate struct.