[backend] Notify host when all players are in
Instead of handling it client side send EventReadyToStart
when all players are in to the host. Host has to click a ready button
which will send EventRequestToStart to the server. The server replies
with WordPhaseStart to all players which changes the client's view.

Rename some of the events to be more expressive.