[backend] Add Quit channels for players
In order to be able to end the listen function for a player
who has quit a game, add quit channel for every player, signal in it on
quit and check the quit channel in the listen func.
The Players.Users map is changed so a user can be deleted by id when
they quit.