[frontend] Delete Ended use Started instead
Reuse the ended interface for displaying results but instead of using a
different Page use processStatus BetweenStages. The end stage is just
BetweenStages with ended flag up. When the stage is BetweenStages
display Start button if the current user is the host and nothing
otherwise. When the flag ended is up display End button for all.